+++ src/test/kotlin/com/github/ekvedaras/classfactoryphpstorm/StateMethodInsideFactoryTest.kt
@@ -6,6 +6,8 @@ import com.github.ekvedaras.classfactoryphpstorm.integration.otherMethods.inspec
 import com.github.ekvedaras.classfactoryphpstorm.integration.otherMethods.inspection.PropertyNotFoundInspection
 import com.github.ekvedaras.classfactoryphpstorm.integration.otherMethods.inspection.PropertyNotFoundInspectionInArrayKeysInDirectlyPassedClosure
 import com.github.ekvedaras.classfactoryphpstorm.integration.otherMethods.inspection.PropertyNotFoundInspectionInAttributesArrayKeys
+import com.github.ekvedaras.classfactoryphpstorm.support.ClassPropertyReference
+import com.intellij.usages.ReadWriteAccessUsageInfo2UsageAdapter
 import com.jetbrains.php.lang.inspections.PhpInspection
 
 internal class StateMethodInsideFactoryTest : EssentialTestCase() {
@@ -38,5 +40,16 @@ internal class StateMethodInsideFactoryTest : EssentialTestCase() {
         assertInspection("nestedStateCalls.php", this.propertyNotFoundInspection())
         assertInspection("nestedStateCalls.php", this.propertyNotFoundInAttributesArrayInspection())
         assertInspection("nestedStateCalls.php", this.propertyNotFoundInArrayKeysInDirectlyPassedClosure())
+
+        val usages = myFixture.testFindUsagesUsingAction("nestedStateCalls.php")
+
+        assertEquals(2, usages.size)
+
+        usages.map { it as ReadWriteAccessUsageInfo2UsageAdapter }.forEach { usage ->
+            assertEquals(ClassPropertyReference::class.java, usage.referenceClass)
+            assertTrue(usage.element?.textMatches("'value'") ?: false)
+            assertTrue(usage.element?.textRange?.startOffset == usage.navigationRange.startOffset - 1)
+            assertTrue(usage.element?.textRange?.endOffset == usage.navigationRange.endOffset + 1)
+        }
     }
 }